backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
certifi==2024.8.30
defusedxml==0.7.1
distro==1.9.0
Django==5.1.3
django-cors-headers==4.6.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
image==1.5.33
jiter==0.7.0
olefile==0.47
openai==1.54.3
pillow==11.0.0
pydantic==2.9.2
pydantic_core==2.23.4
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.1
tqdm==4.67.0
typing_extensions==4.12.2
